refactor(refund-tables) structure and add new form component
