+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/refund-fees/_components/table-data.tsx
@@ -63,6 +63,7 @@ const refundFeeHeadersColumns = (
           ],
         },
       },
+      // excludeColumns: ["isActive"],
     },
   );
