+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/refund-tables/[id]/page.tsx
@@ -2,7 +2,7 @@ import { notFound } from "next/navigation";
 import { getRefundTableHeadersByIdApi } from "src/actions/unirefund/ContractService/action";
 import { isUnauthorized } from "src/utils/page-policy/page-policy";
 import { getResourceData } from "src/language-data/unirefund/ContractService";
-import Form from "./form";
+import Form from "./_components/form";
 
 export default async function Page({
   params,
