+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/refund-tables/new/_components/form.tsx
@@ -1,0 +1,3 @@
+export default function RefundTableHeaderCreateForm() {
+  return <div>Form</div>;
+}
