+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/refund-tables/new/page.tsx
@@ -1,0 +1,5 @@
+import RefundTableHeaderCreateForm from "./_components/form";
+
+export default function Page() {
+  return <RefundTableHeaderCreateForm />;
+}
