+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/refund-tables/page.tsx
@@ -5,7 +5,7 @@ import { notFound } from "next/navigation";
 import { getResourceData } from "src/language-data/unirefund/ContractService";
 import { isUnauthorized } from "src/utils/page-policy/page-policy";
 import { getRefundTableHeadersApi } from "../../../../../../../actions/unirefund/ContractService/action";
-import RefundTable from "./table";
+import RefundTable from "./_components/table";
 
 export default async function Page(props: {
   params: { lang: string };
